[crypto] Add testdata::rsa
