+++ src/crypto/testdata/mod.rs
@@ -8,6 +8,11 @@
 //! only. In particular, we keep the private key around to make it possible to
 //! re-sign test data to reduce the brittleness of tests.
 
+use crate::crypto::ring;
+use crate::crypto::rsa::Builder as _;
+use crate::crypto::rsa::Keypair as _;
+use crate::crypto::rsa::SignerBuilder as _;
+
 /// A plaintext string.
 pub const PLAIN_TEXT: &[u8] = include_bytes!("plain.txt");
 
@@ -32,3 +37,13 @@ pub const RSA_2048_PRIV_PKCS8: &[u8] =
 /// The signature is in PKCS v1.5 format.
 pub const RSA_2048_SHA256_SIG_PKCS1: &[u8] =
     include_bytes!("rsa_2048_sha256_sig.pk1");
+
+/// Generates an RSA engine and signer out of test-only data.
+pub fn rsa() -> (ring::rsa::Engine, ring::rsa::Signer) {
+    let keypair = ring::rsa::Keypair::from_pkcs8(RSA_2048_PRIV_PKCS8).unwrap();
+    let pub_key = keypair.public();
+    let rsa_builder = ring::rsa::Builder::new();
+    let rsa = rsa_builder.new_engine(pub_key).unwrap();
+    let signer = rsa_builder.new_signer(keypair).unwrap();
+    (rsa, signer)
+}
